fix: sanitize preview middleware configuration if necessary
