+++ .changeset/brown-suns-jog.md
@@ -1,0 +1,5 @@
+---
+'@sap-ux/app-config-writer': patch
+---
+
+fix: sanitize preview middleware configuration if necessary
